+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Fixed
 
 - [#176](https://github.com/hyperf-cloud/hyperf/pull/176) Fixed TypeError: Return value of LengthAwarePaginator::nextPageUrl() must be of the type string or null, none returned.
+- [#188](https://github.com/hyperf-cloud/hyperf/pull/188) Fixed proxy of guzzle client does not work expected.
 
 # v1.0.4 - 2019-07-08